Add timeouts for answers validation and upgrade
In the case that the upgrade process took more than 5 minutes, the
upgrade could be marked as failed and would be terminated. On loaded
machines this could kill valid but slow upgrades. To resolve this, the
timeout value used for downloading was extended to the installation
process.

This fixes issues #1 and #2