Ensure the module doesn't run on Puppet masters.
A common use case for PE upgrades is:
  - User upgrades the master to the latest PE release
  - User allows the master to run for awhile looking for any problems
  - Potentially much later the user updates the PE Upgrade module
    variables to the latest version
  - Puppet manages agent upgrades across the cluster

A problem can arise when the PE Upgrade module is also managing the
master (as is commonly the case with dashboard groups). The user will
upgrade the master, the PE upgrade modules variable still reflect the
old version, the PE Upgrade module will downgrade the master before the
user updates the variables.

The PE Upgrade module should not interact with the master.