Add explicit fail and force_upgrade for masters
Per (pe_upgrade-#7) it is possible for a master to forcibly downgrade
itself if the master is upgraded but has old data referring to an old
version. The patch supplied would disable upgrading a master but had the
unfortunate side effect of silently refusing to upgrade. This adds an
explicit failure to prevent masters from being modified by accident.