Remove pe_version facts, add stdlib dependency
When both pe_upgrade and stdlib are installed, the pe_upgrade facts win out
(not sure what pluginsync does with two facter filenames that are the same, but
it appears that the first in lexigraphic order is preferred and once the file
is there it isn't replaced), but are missing a key fact is_pe. This commit
addresses that problem by removing the facts from pe_upgrade and adding a
dependency on stdlib of 2.5.1 or higher (where the pe_version facts will be
available).