Purge PE upgrade staging directory.
Over a series of upgrades the staging directory will accumulate PE
tarballs and directories. This change will purge the module's staging
directory when the PE versions are in agreement.

We purge the directory recursively to get the untarred directories as
well as the tarballs. We disable backups because we don't want to send
anything to a filebucket.